Equip Issue resolve, IInteractable distance, editor floating
Took a few minutes out of my day to fix a few things that were on my list. Equipping a primary weapon with a both type weapon in hand will now move it to secondary slot if its empty, instead of automatically throwing it. Interact check distance was altered to be about the same length for first and third person, instead of first being considerably further. Added an in game mode check for the update to stop flying animation from playing while idle in editor.